app/requirements.txt
Flask
logstash-formatter
certifi==2024.8.30
charset-normalizer==3.3.2
coverage==7.6.1
Deprecated==1.2.14
idna==3.10
importlib_resources==6.4.5
jaeger-client==4.8.0
limits==3.13.0
opentelemetry-api==1.7.1
opentelemetry-exporter-jaeger-thrift==1.7.1
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-celery==0.26b1
opentelemetry-instrumentation-falcon==0.26b1
opentelemetry-instrumentation-logging==0.26b1
opentelemetry-instrumentation-wsgi==0.26b1
opentelemetry-propagator-jaeger==1.7.1
opentelemetry-sdk==1.7.1
opentelemetry-semantic-conventions==0.26b1
opentelemetry-util-http==0.26b1
opentracing==2.4.0
packaging==24.1
pylogbeat==2.0.1
python-logstash-async==2.3.0
requests==2.32.3
setuptools==75.1.0
six==1.16.0
threadloop==1.0.2
thrift==0.21.0
tornado==6.4.1
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
